+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.27.1] - 2022-01-03
+
+### Fixed
+
+- Fixed a bug in the RTK base station handling that sometimes resulted in an
+  unhandled error message when the RTK base station was not sending valid
+  messages at all.
+
 ## [1.27.0] - 2021-12-29
 
 ### Added